Bullet Parsing and Database saving working on Create Ammo Screen

+++ app/src/main/java/jant/gunreloads/app/sql/model/Manufacturer.java
@@ -12,7 +12,7 @@ public class Manufacturer extends PersistanceObject {
         this.name = name;
     }
 
-    public Manufacturer(int id, String name) {
+    public Manufacturer(long id, String name) {
         this.id = id;
         this.name = name;
     }
